refactor: extract DiagnosticsEventName.swift from DiagnosticsTypes.swift
Splits the `DiagnosticsEventName` value type (RawRepresentable +
ExpressibleByStringLiteral) out of DiagnosticsTypes.swift.

Byte-identical move of lines 38-52. New file needs only Foundation
(no OSLog).

DiagnosticsTypes.swift: 92 → 76 lines. Tests: 667.
